misc: make the code better
